+++ agents/util.go
@@ -99,7 +99,6 @@ func Gorequest(o core.Options) *gorequest.SuperAgent {
 	return gorequest.New().
 		Proxy(o.Proxy).
 		Timeout(time.Duration(o.HTTPTimeout)*time.Millisecond).
-		SetDebug(o.Debug).
 		TLSClientConfig(&tls.Config{InsecureSkipVerify: true})
 }