+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Added:
+- Supported optional headers
+
+## Changed:
+- Updated colors/styles of report tags
+
 ## [1.9.0-shelld3v]
 
 ### Added:
@@ -11,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Able to filter hosts by string in response
 
 ### Changed:
-- Refactor finger-printing feature
+- Refactoring of finger-printing feature
 - Removed prototype pollution check
 - Updated request headers